Move OpenAIMaxTokensRewriteTransform into protocol/transform/ package
The transform only depends on the OpenAI SDK and the ops package — no
server-domain types — so it belongs alongside BaseTransform /
ConsistencyTransform / VendorTransform in internal/protocol/transform/,
not internal/server/. internal/server/transform_*.go remains the home for
transforms that need server-domain types (ThinkingModeTransform reading
*typ.ScenarioConfig, etc.).

- Move transform_openai_max_tokens_rewrite.go ->
  protocol/transform/openai_max_tokens_rewrite.go
- Move companion test file
- Package decl transform; drop the protocoltransform alias since the
  type now lives in that package
- rule_flags.go: qualify constructor as transform.NewOpenAIMaxTokensRewriteTransform
- rule_flags_test.go: qualify type assertion as
  *transform.OpenAIMaxTokensRewriteTransform
- Design doc table now distinguishes protocol-layer Transforms from
  server-domain Transforms and reflects the new path
